Fix bugs from new settings
