+++ crates/remote_server/src/unix.rs
@@ -1,17 +1,23 @@
 use crate::HeadlessProject;
 use anyhow::{anyhow, Context, Result};
-use fs::RealFs;
+use client::ProxySettings;
+use fs::{Fs, RealFs};
 use futures::channel::mpsc;
 use futures::{select, select_biased, AsyncRead, AsyncWrite, AsyncWriteExt, FutureExt, SinkExt};
-use gpui::{AppContext, Context as _};
+use gpui::{AppContext, Context as _, ModelContext, UpdateGlobal as _};
+use http_client::{read_proxy_from_env, Uri};
+use node_runtime::{NodeBinaryOptions, NodeRuntime};
 use paths::logs_dir;
+use project::project_settings::ProjectSettings;
 use remote::proxy::ProxyLaunchError;
 use remote::ssh_session::ChannelClient;
 use remote::{
     json_log::LogRecord,
     protocol::{read_message, write_message},
 };
-use rpc::proto::Envelope;
+use reqwest_client::ReqwestClient;
+use rpc::proto::{self, Envelope, SSH_PROJECT_ID};
+use settings::{watch_config_file, Settings, SettingsStore};
 use smol::channel::{Receiver, Sender};
 use smol::io::AsyncReadExt;
 
@@ -23,6 +29,7 @@ use std::{
     path::{Path, PathBuf},
     sync::Arc,
 };
+use util::ResultExt;
 
 fn init_logging_proxy() {
     env_logger::builder()
@@ -316,13 +323,28 @@ pub fn execute_run(
         let session = start_server(listeners, log_rx, cx);
 
         let project = cx.new_model(|cx| {
-            HeadlessProject::new(session, Arc::new(RealFs::new(Default::default(), None)), cx)
-        });
+            let fs = Arc::new(RealFs::new(Default::default(), None));
+            let node_settings_rx = initialize_settings(session.clone(), fs.clone(), cx);
 
-        cx.on_app_quit(|_cx| async {
-            log::info!("remote server is gracefully shutting down");
-        })
-        .detach();
+            let proxy_url = read_proxy_settings(cx);
+
+            let http_client = Arc::new(
+                ReqwestClient::proxy_and_user_agent(
+                    proxy_url,
+                    &format!(
+                        "Zed-Server/{} ({}; {})",
+                        env!("CARGO_PKG_VERSION"),
+                        std::env::consts::OS,
+                        std::env::consts::ARCH
+                    ),
+                )
+                .expect("Could not start HTTP client"),
+            );
+
+            let node_runtime = NodeRuntime::new(http_client.clone(), node_settings_rx);
+
+            HeadlessProject::new(session, fs, http_client, node_runtime, cx)
+        });
 
         mem::forget(project);
     });
@@ -370,7 +392,6 @@ pub fn execute_proxy(identifier: String, is_reconnecting: bool) -> Result<()> {
 
     let server_pid = check_pid_file(&server_paths.pid_file)?;
     let server_running = server_pid.is_some();
-    let mut child_task = None;
     if is_reconnecting {
         if !server_running {
             log::error!("attempted to reconnect, but no server running");
@@ -382,7 +403,7 @@ pub fn execute_proxy(identifier: String, is_reconnecting: bool) -> Result<()> {
             kill_running_server(pid, &server_paths)?;
         }
 
-        child_task = Some(spawn_server(&server_paths)?);
+        spawn_server(&server_paths)?;
     };
 
     let stdin_task = smol::spawn(async move {
@@ -416,18 +437,6 @@ pub fn execute_proxy(identifier: String, is_reconnecting: bool) -> Result<()> {
             }
         }
     });
-
-    if let Some(mut child) = child_task {
-        smol::spawn(async move {
-            log::info!(
-                "Server process finished with exit code {:?}",
-                child.status().await
-            )
-        })
-        .detach();
-    } else {
-        log::info!("Reconnected to existing server process, no child process to report.")
-    }
 
     if let Err(forwarding_result) = smol::block_on(async move {
         futures::select! {
@@ -465,7 +474,7 @@ fn kill_running_server(pid: u32, paths: &ServerPaths) -> Result<()> {
     Ok(())
 }
 
-fn spawn_server(paths: &ServerPaths) -> Result<smol::process::Child> {
+fn spawn_server(paths: &ServerPaths) -> Result<()> {
     if paths.stdin_socket.exists() {
         std::fs::remove_file(&paths.stdin_socket)?;
     }
@@ -509,7 +518,7 @@ fn spawn_server(paths: &ServerPaths) -> Result<smol::process::Child> {
         total_time_waited
     );
 
-    Ok(server_process)
+    Ok(())
 }
 
 fn check_pid_file(path: &Path) -> Result<Option<u32>> {
@@ -580,4 +589,119 @@ async fn write_size_prefixed_buffer<S: AsyncWrite + Unpin>(
     stream.write_all(len.to_le_bytes().as_slice()).await?;
     stream.write_all(buffer).await?;
     Ok(())
+}
+
+fn initialize_settings(
+    session: Arc<ChannelClient>,
+    fs: Arc<dyn Fs>,
+    cx: &mut AppContext,
+) -> async_watch::Receiver<Option<NodeBinaryOptions>> {
+    let user_settings_file_rx = watch_config_file(
+        &cx.background_executor(),
+        fs,
+        paths::settings_file().clone(),
+    );
+
+    handle_settings_file_changes(user_settings_file_rx, cx, {
+        let session = session.clone();
+        move |err, _cx| {
+            if let Some(e) = err {
+                log::info!("Server settings failed to change: {}", e);
+
+                session
+                    .send(proto::Toast {
+                        project_id: SSH_PROJECT_ID,
+                        notification_id: "server-settings-failed".to_string(),
+                        message: format!(
+                            "Error in settings on remote host {:?}: {}",
+                            paths::settings_file(),
+                            e
+                        ),
+                    })
+                    .log_err();
+            } else {
+                session
+                    .send(proto::HideToast {
+                        project_id: SSH_PROJECT_ID,
+                        notification_id: "server-settings-failed".to_string(),
+                    })
+                    .log_err();
+            }
+        }
+    });
+
+    let (tx, rx) = async_watch::channel(None);
+    cx.observe_global::<SettingsStore>(move |cx| {
+        let settings = &ProjectSettings::get_global(cx).node;
+        log::info!("Got new node settings: {:?}", settings);
+        let options = NodeBinaryOptions {
+            allow_path_lookup: !settings.ignore_system_version.unwrap_or_default(),
+            // TODO: Implement this setting
+            allow_binary_download: true,
+            use_paths: settings.path.as_ref().map(|node_path| {
+                let node_path = PathBuf::from(shellexpand::tilde(node_path).as_ref());
+                let npm_path = settings
+                    .npm_path
+                    .as_ref()
+                    .map(|path| PathBuf::from(shellexpand::tilde(&path).as_ref()));
+                (
+                    node_path.clone(),
+                    npm_path.unwrap_or_else(|| {
+                        let base_path = PathBuf::new();
+                        node_path.parent().unwrap_or(&base_path).join("npm")
+                    }),
+                )
+            }),
+        };
+        tx.send(Some(options)).log_err();
+    })
+    .detach();
+
+    rx
+}
+
+pub fn handle_settings_file_changes(
+    mut server_settings_file: mpsc::UnboundedReceiver<String>,
+    cx: &mut AppContext,
+    settings_changed: impl Fn(Option<anyhow::Error>, &mut AppContext) + 'static,
+) {
+    let server_settings_content = cx
+        .background_executor()
+        .block(server_settings_file.next())
+        .unwrap();
+    SettingsStore::update_global(cx, |store, cx| {
+        store
+            .set_server_settings(&server_settings_content, cx)
+            .log_err();
+    });
+    cx.spawn(move |cx| async move {
+        while let Some(server_settings_content) = server_settings_file.next().await {
+            let result = cx.update_global(|store: &mut SettingsStore, cx| {
+                let result = store.set_server_settings(&server_settings_content, cx);
+                if let Err(err) = &result {
+                    log::error!("Failed to load server settings: {err}");
+                }
+                settings_changed(result.err(), cx);
+                cx.refresh();
+            });
+            if result.is_err() {
+                break; // App dropped
+            }
+        }
+    })
+    .detach();
+}
+
+fn read_proxy_settings(cx: &mut ModelContext<'_, HeadlessProject>) -> Option<Uri> {
+    let proxy_str = ProxySettings::get_global(cx).proxy.to_owned();
+    let proxy_url = proxy_str
+        .as_ref()
+        .and_then(|input: &String| {
+            input
+                .parse::<Uri>()
+                .inspect_err(|e| log::error!("Error parsing proxy settings: {}", e))
+                .ok()
+        })
+        .or_else(read_proxy_from_env);
+    proxy_url
 }
